Create fake data and seed it to database
