+++ .gitignore
@@ -24,3 +24,4 @@ coverage
 build/
 dist/
 
+data/
